+++ web/Drug.jsp
@@ -26,6 +26,16 @@
 <!--===============================================================================================-->
     </head>
     <body>
+        <%               
+            String user = (String)session.getAttribute("username");
+            String role = (String)session.getAttribute("role");
+            if(user != null) {
+                //out.println("Welcome, " + user);
+            }
+            else {
+                response.sendRedirect("login.jsp");
+            }
+        %>
         <div class="row">             
             
              <div class="col-md-12 col-sm-12 col-xs-12">
@@ -57,7 +67,9 @@
                       %>
                       Drug Record.
                     </p>
-                    <a role="button" class="btn btn-default col-sm-1" href="AddDrug.jsp?status=Add&id=">Add..</a><br><br>
+                    <%if(role.equals("1")){
+                        %><a role="button" class="btn btn-default col-sm-1" href="AddDrug.jsp?status=Add&id=">Add..</a><br><br><%
+                    }%>
                     <table id="datatable-fixed-header" class="table table-striped table-bordered">
                       <thead>
                         <tr>
@@ -67,7 +79,9 @@
                           <th>Stock</th>
                           <th>Drug Exp</th>
                           <th>Price</th>
-                          <th>Action</th>
+                          <%if(role.equals("1")){
+                              %><th>Action</th><%
+                          }%>
                         </tr>
                       </thead>
                       <tbody>
@@ -91,10 +105,12 @@
                             out.println("<td>"+Stock+"</td>");
                             out.println("<td>"+DrugExp+"</td>");
                             out.println("<td>"+Price+"</td>");
-                            out.println("<td>");
-                            out.println("<a role='button' class='btn btn-info' href='AddDrug.jsp?status=Update&id="+idDrug+"' >Modify </a>");
-                            out.println("<a role='button' class='btn btn-danger' href='delete_book.jsp?id="+idDrug+"'>Delete </a>");
-                            out.println("</td>");
+                            if(role.equals("1")){
+                                out.println("<td>");
+                                out.println("<a role='button' class='btn btn-info' href='AddDrug.jsp?status=Update&id="+idDrug+"' >Modify </a>");
+                                out.println("<a role='button' class='btn btn-danger' href='delete_book.jsp?id="+idDrug+"'>Delete </a>");
+                                out.println("</td>");
+                            }
                             out.println("</tr>");
                         }
                     }
